requirements.txt
-f https://download.pytorch.org/whl/torch_stable.html
Flask==3.0.2
gunicorn==22.0.0
Werkzeug==3.0.1
Jinja2==3.1.3
MarkupSafe==2.1.5
itsdangerous==2.1.2
click==8.1.7
requests==2.31.0
torch==2.3.0+cpu
torchvision==0.18.0+cpu
Pillow==10.2.0
boto3
python-dotenv==0.19.1
